requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
environs==11.0.0
inflection==0.5.1
marshmallow==3.21.3
packaging==24.1
pillow==10.3.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.7.0
